requirements.txt
pillow
argparse